+++ ibc/errors.go
@@ -1,0 +1,8 @@
+package ibc
+
+import "errors"
+
+var (
+	ErrNoIBCVoucherDenom  = errors.New("denom is not an IBC voucher")
+	ErrDenomTraceNotFound = errors.New("denom trace not found")
+)